update table schema for pgsql

+++ OracleSpatialWeb/sql/PostgreSQL/DDL/flickr_photo.sql
@@ -10,12 +10,12 @@ CREATE TABLE flickr_photo
   latitude numeric(15,0) NOT NULL,
   taken_date date,
   upload_date date,
-  viewed numeric,
+  viewed integer,
   title character varying(255),
   smallurl character varying(100),
   place_id character varying(20),
   woe_id character varying(20),
-  accuracy numeric,
+  accuracy integer,
   CONSTRAINT flickr_photo_pk PRIMARY KEY (photo_id),
   CONSTRAINT flickr_photo_fk FOREIGN KEY (user_id)
       REFERENCES flickr_people (user_id) MATCH SIMPLE
